Format errors and move default options into config.json

+++ src/tokens/mod.rs
@@ -7,7 +7,7 @@ use assignment::Assignment;
 
 #[derive(Debug)]
 pub struct CodeLocation {
-  pub line: u16,
+  pub line: usize,
   pub column: usize,
 }
